fix: comprehensive code quality improvements and security fixes
This commit addresses multiple code quality, security, and compatibility issues across 54 files:

Security Fixes:
- Remove API key length exposure from error messages in bluefin_client.py
- Update langchain-community to v0.2.0+ to fix known CVE vulnerabilities
- Remove sensitive JWT token preview from debug logs in market.py
- Improve error messages to avoid exposing sensitive configuration details

Code Quality Improvements:
- Apply Ruff formatting to improve code readability and consistency
- Fix logging format errors by replacing f-strings with % formatting in logging calls
- Resolve critical type errors in market_context.py and fee_calculator.py
- Fix undefined name errors that could cause runtime failures
- Remove dead code (unused exc_tb parameter in bluefin_service_client.py)
- Clean up duplicate metadata and dependencies in pyproject.toml

Test Suite Fixes:
- Update test models for Pydantic v2 compatibility
- Add missing required fields (symbol, timestamp) to Position model creation
- Fix mock configuration issues in test_position_manager.py
- Improve test reliability and reduce flakiness

Configuration and Documentation:
- Add configuration validation scripts for comprehensive setup verification
- Update Docker permission fixes documentation
- Improve emergency reset instructions
- Add configuration security recommendations

Performance and Reliability:
- Optimize trade aggregation logic in market data handlers
- Improve WebSocket connection stability in bluefin_websocket.py
- Fix race conditions in order management systems
- Enhance error handling in fee calculation modules

The changes improve security, code quality, and test reliability without affecting core functionality.
All modifications maintain backward compatibility while addressing critical issues.

+++ bot/fee_calculator.py
@@ -166,7 +166,7 @@ class FeeCalculator:
             return fees
 
     def _create_zero_fee_result(
-        self, trade_action: TradeAction, size_pct: int | None = None
+        self, trade_action: TradeAction, size_pct: float | None = None
     ) -> tuple[TradeAction, TradeFees]:
         """Helper method to create a zero fee result with adjusted trade action."""
         adjusted_action = trade_action.copy()
